ENH: Adding the support of GEOSAT-2 constellation #59
